Communication with Database
Was able to establish a connection with the database few days ago after
fighting with dependencies and establishing connections when not running
the website from main().  Will work on pulling info from database for
both instructor and student site, clean up the layout of both sites, and
making a class to handle communication with the database.