Save Feature active
We are now able to save the selfAssessment when logged on with a
student.  (It works with my login)  We need to work on adding students
to the course/group/ROLE/user table etc.   Maybe see about getting
started with the query/transaction that will handle that when the
instructor adds the student to the course/group