Able to add members to group by UCA.
It has taken about 8 hours, but we can add members to groups now.  If
their UCA is not in the database, instructor is prompted to enter a name
for each UCA that doesn't exit.  They are then inserted into the user
table, the projectmanager or dev table(determined by role button), and
added to the groupMemberList